fix inconsistency with currency types

+++ src/player.cpp
@@ -5,7 +5,7 @@ double Player::get_balance()
     return balance;
 }
 
-bool Player::set_balance(int amount)
+bool Player::set_balance(double amount)
 {
     if (amount > 0)
     {
