+++ src/player.h
@@ -6,7 +6,7 @@ public:
     // Getters & Setter
     double get_balance();
 
-    bool set_balance(int amount);
+    bool set_balance(double amount);
 
 private:
     double balance = 100;
